requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
Flask==3.0.3
pandas==2.2.2
psutil==5.9.1
pytest==6.2.4
pluggy==0.13.1
pywinauto==0.6.8
PyYAML==5.3.1
requests==2.32.3
selenium==4.23.1
xlrd==2.0.1
xlwt==1.3.0
XMind==1.2.0
xmind2testcase==1.5.0
webdriver-manager==4.0.2
python-dateutil~=2.9.0.post0
pillow~=11.0.0
pytesseract~=0.3.13
openpyxl==3.1.5
pyodbc==5.2.0